Update ICD to Version 1.0 (07.08.2024) Update SEDAP Express Framework project to version 1.0.0 Update misc sample projects

+++ SEDAPExpress/src/main/java/de/bundeswehr/mese/sedapexpress/messages/EMISSION.java
@@ -269,7 +269,7 @@ public class EMISSION extends SEDAPExpressMessage {
      * @param sidc
      * @param comment
      */
-    public EMISSION(Short number, Long time, String sender, Character classification, Boolean acknowledgement, String mac, String emissionID, Boolean deleteFlag, Double sensorLatitude, Double sensorLongitude, Double sensorAltitude,
+    public EMISSION(Byte number, Long time, String sender, Character classification, Boolean acknowledgement, String mac, String emissionID, Boolean deleteFlag, Double sensorLatitude, Double sensorLongitude, Double sensorAltitude,
 	    Double emitterLatitude, Double emitterLongitude, Double emitterAltitude, Double bearing, List<Double> frequency, Double bandwidth, Double power, Integer freqAgility, Integer prfAgility, Integer function, Integer spotNumber, char[] sidc,
 	    String comment) {
