+++ SEDAPExpress/src/main/java/de/bundeswehr/mese/sedapexpress/messages/HEARTBEAT.java
@@ -51,7 +51,7 @@ public class HEARTBEAT extends SEDAPExpressMessage {
      * @param acknowledgement
      * @param mac
      */
-    public HEARTBEAT(Short number, Long time, String sender, Character classification, Boolean acknowledgement, String mac) {
+    public HEARTBEAT(Byte number, Long time, String sender, Character classification, Boolean acknowledgement, String mac) {
 
 	super(number, time, sender, classification, acknowledgement, mac);
 
@@ -68,7 +68,7 @@ public class HEARTBEAT extends SEDAPExpressMessage {
      * @param mac
      * @param recipient
      */
-    public HEARTBEAT(Short number, Long time, String sender, Character classification, Boolean acknowledgement, String mac, String recipient) {
+    public HEARTBEAT(Byte number, Long time, String sender, Character classification, Boolean acknowledgement, String mac, String recipient) {
 
 	super(number, time, sender, classification, acknowledgement, mac);
