+++ SEDAPExpress/src/main/java/de/bundeswehr/mese/sedapexpress/messages/OWNUNIT.java
@@ -147,7 +147,7 @@ public class OWNUNIT extends SEDAPExpressMessage {
      * @param name
      * @param sidc
      */
-    public OWNUNIT(Short number, Long time, String sender, Character classification, Boolean acknowledgement, String mac,
+    public OWNUNIT(Byte number, Long time, String sender, Character classification, Boolean acknowledgement, String mac,
 	    Double latitude, Double longitude, Double altitude, Double speed, Double course, Double heading,
 	    Double roll, Double pitch, String name, char[] sidc) {
